Majorly revamped the file handling connections
build time reduced by a factor of 4. Commands testing tomorrow
