+++ org.eclipse.ice.commands/src/test/java/org/eclipse/ice/tests/commands/RemoteCopyFileCommandTest.java
@@ -16,15 +16,11 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 
 import org.eclipse.ice.commands.CommandStatus;
-import org.eclipse.ice.commands.Connection;
-import org.eclipse.ice.commands.ConnectionAuthorizationHandler;
-import org.eclipse.ice.commands.ConnectionAuthorizationHandlerFactory;
-import org.eclipse.ice.commands.ConnectionConfiguration;
 import org.eclipse.ice.commands.ConnectionManager;
 import org.eclipse.ice.commands.ConnectionManagerFactory;
 import org.eclipse.ice.commands.RemoteCopyFileCommand;
 import org.junit.AfterClass;
-import org.junit.Before;
+import org.junit.BeforeClass;
 import org.junit.Test;
 
 import com.jcraft.jsch.ChannelSftp;
@@ -49,50 +45,21 @@ public class RemoteCopyFileCommandTest {
 	String dest = null;
 
 	/**
-	 * A connection with which to test
-	 */
-	ConnectionConfiguration connectionConfig = new ConnectionConfiguration();
-
-	/**
-	 * A connection to move files with
-	 */
-	Connection dummyConnection = null;
-
-	/**
-	 * A connection manager for opening the connection to test with
-	 */
-	ConnectionManager manager = new ConnectionManager();
-
-	/**
-	 * Make a IFileHandlerFactoryTest to take advantage of much of the code which
+	 * Make a RemoteFileHandlerTest to take advantage of much of the code which
 	 * makes/deletes local/remote files
 	 */
-	IFileHandlerFactoryTest factory = new IFileHandlerFactoryTest();
+	RemoteFileHandlerTest handlerTest = new RemoteFileHandlerTest();
 
 	/**
+	 * This function sets up the dummy connection for file transferring
+	 * 
 	 * @throws java.lang.Exception
 	 */
-	@Before
-	public void setUp() throws Exception {
-
-		// Set the connection configuration to a dummy remote connection
-		// Make the connection configuration
-		// Get a factory which determines the type of authorization
-		ConnectionAuthorizationHandlerFactory authFactory = new ConnectionAuthorizationHandlerFactory();
-		// Request a ConnectionAuthorization of type text file which contains the
-		// credentials
-		ConnectionAuthorizationHandler auth = authFactory.getConnectionAuthorizationHandler("text",
-				"/tmp/ice-remote-creds.txt");
-		// Set it
-		connectionConfig.setAuthorization(auth);
-		connectionConfig.setName("dummyConnection");
-		connectionConfig.setDeleteWorkingDirectory(false);
-
-		// open the connection and set it so that the factory can create the
-		// remote files
-		dummyConnection = manager.openConnection(connectionConfig);
-		factory.setConnection(dummyConnection);
-
+	@BeforeClass
+	public static void setUpBeforeClass() throws Exception {
+		// Set up the connection using the code already established in
+		// RemoteFileHandlerTest
+		RemoteFileHandlerTest.setUpBeforeClass();
 	}
 
 	/**
@@ -113,16 +80,16 @@ public class RemoteCopyFileCommandTest {
 	@Test
 	public void testRemoteCopyFileCommand() throws Exception {
 
-		factory.createRemoteSource();
-		factory.createRemoteDestination();
-		source = factory.getSource();
-		dest = factory.getDestination();
+		handlerTest.createRemoteSource();
+		handlerTest.createRemoteDestination();
+		source = handlerTest.getSource();
+		dest = handlerTest.getDestination();
 
 		RemoteCopyFileCommand command = new RemoteCopyFileCommand();
 		// These functions are nominally handled by the FileHandler. But, when testing
 		// this class alone, we need to set them individually
 		command.setCopyType(3);
-		command.setConnection(dummyConnection);
+		command.setConnection(handlerTest.getConnection());
 		command.setConfiguration(source, dest);
 		CommandStatus status = command.execute();
 
@@ -134,8 +101,8 @@ public class RemoteCopyFileCommandTest {
 		assert (remotePathExists());
 
 		// Delete the temporary files that were created to test
-		factory.deleteRemoteSource();
-		factory.deleteRemoteDestination();
+		handlerTest.deleteRemoteSource();
+		handlerTest.deleteRemoteDestination();
 
 	}
 
@@ -148,18 +115,18 @@ public class RemoteCopyFileCommandTest {
 	@Test
 	public void testRemoteCopyFileCommandDownload() throws Exception {
 		// Create a remote source to download
-		factory.createRemoteSource();
-		source = factory.getSource();
+		handlerTest.createRemoteSource();
+		source = handlerTest.getSource();
 
 		// Create a local destination to put that source
-		factory.createLocalDestination();
-		dest = factory.getDestination();
+		handlerTest.createLocalDestination();
+		dest = handlerTest.getDestination();
 
 		RemoteCopyFileCommand command = new RemoteCopyFileCommand();
 		// These functions are nominally handled by the FileHandler. But, when testing
 		// this class alone, we need to set them individually
 		command.setCopyType(2);
-		command.setConnection(dummyConnection);
+		command.setConnection(handlerTest.getConnection());
 		command.setConfiguration(source, dest);
 		CommandStatus status = command.execute();
 
@@ -171,8 +138,8 @@ public class RemoteCopyFileCommandTest {
 		assert (localPathExists());
 
 		// Delete the temporary files that were created to test
-		factory.deleteRemoteSource();
-		factory.deleteLocalDestination();
+		handlerTest.deleteRemoteSource();
+		handlerTest.deleteLocalDestination();
 
 	}
 
@@ -184,17 +151,17 @@ public class RemoteCopyFileCommandTest {
 	public void testRemoteCopyFileCommandUpload() throws Exception {
 
 		// Create a local source file to move
-		factory.createLocalSource();
-		source = factory.getSource();
+		handlerTest.createLocalSource();
+		source = handlerTest.getSource();
 		// Create a remote destination to move it to
-		factory.createRemoteDestination();
-		dest = factory.getDestination();
+		handlerTest.createRemoteDestination();
+		dest = handlerTest.getDestination();
 
 		// Make a command and execute it
 		RemoteCopyFileCommand command = new RemoteCopyFileCommand();
 		// These functions are nominally handled by the FileHandler. But, when testing
 		// this class alone, we need to set them individually
-		command.setConnection(dummyConnection);
+		command.setConnection(handlerTest.getConnection());
 		command.setCopyType(1);
 		command.setConfiguration(source, dest);
 
@@ -210,8 +177,8 @@ public class RemoteCopyFileCommandTest {
 			e.printStackTrace();
 		}
 
-		factory.deleteLocalSource();
-		factory.deleteRemoteDestination();
+		handlerTest.deleteLocalSource();
+		handlerTest.deleteRemoteDestination();
 	}
 
 	/**
@@ -232,9 +199,8 @@ public class RemoteCopyFileCommandTest {
 	 */
 	public boolean remotePathExists() throws Exception {
 
-		// Connect the channel from the connection
-		ChannelSftp sftpChannel = (ChannelSftp) dummyConnection.getSession().openChannel("sftp");
-		sftpChannel.connect();
+		// Get the connected channel from the connection
+		ChannelSftp sftpChannel = (ChannelSftp) handlerTest.getConnection().getChannel();
 
 		try {
 			sftpChannel.lstat(dest);
@@ -243,6 +209,7 @@ public class RemoteCopyFileCommandTest {
 			return false;
 		}
 
+		// So if we get here, then the file is there
 		return true;
 	}
